moved all factoid stuff into its own package
