+++ src/main/java/com/fossgalaxy/pircbotx/commands/factoid/FactoidModel.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License along with
  * uc_pircbotx. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.fossgalaxy.pircbotx.data.db;
+package com.fossgalaxy.pircbotx.commands.factoid;
 
 import com.google.inject.Inject;
 import org.slf4j.Logger;
