+++ src/main/java/com/fossgalaxy/pircbotx/data/db/DatabaseModule.java
@@ -1,5 +1,6 @@
 package com.fossgalaxy.pircbotx.data.db;
 
+import com.fossgalaxy.pircbotx.commands.factoid.FactoidModel;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import org.slf4j.Logger;
